Deterministic sort order between shaders
This implements a deterministic sort order between shaders. Otherwise, the batches get sorted by Shader.Id (which is non-deterministic and determined only by the creation order of the shaders). It is important that some shaders (like stencil lines) must be done after the corresonding mesh shader (Phong/Gourad etc). This was sometimes getting broken depending on which shader gets constructed first. Now, the sort order is based on the 'SortCode' field that is provided in the index.txt manifest of all shaders.